Refactor hustle market config builders
